fix(check): detect Nuxt2 src plugin injections
